# "fix" for the sequence file performance