1. add our own view::istream that is faster than std::istreambuf_iterators
2. conditionally remove usage of most views from sequence file fasta
  * not all views seem to be bad in all situations, but view::filter and view::join seems to really kill the performance currently

This also increased the size of the benchmarked FastA-file a little, because the previous one was too small. That's why the numbers for SeqAn2 changed, too.

|          | SeqAn3 before | SeqAn3 after | SeqAn2 before | SeqAn2 after |
|--------|----------------|-------------|---------------|--------------|
| read  GCC7 |  139.474M/s | 454.196M/s | 383.446M/s | 495.321M/s |
| read GCC9 | 119.112M/s  | 510.165M/s | 446.884M/s | 553.42M/s |
| write GCC7 | 164.803M/s | 331.439M/s | 109.172M/s | 109.186M/s |
| write GCC9 | 180.341M/s | 302.799M/s | 113.784M/s | 109.434M/s |